+++ helpers/python.ts
@@ -49,6 +49,17 @@ const getAdditionalDependencies = (
       });
       break;
     }
+    case "milvus": {
+      dependencies.push({
+        name: "llama-index-vector-stores-milvus",
+        version: "^0.1.6",
+      });
+      dependencies.push({
+        name: "pymilvus",
+        version: "2.3.7",
+      });
+      break;
+    }
   }
 
   // Add data source dependencies