+++ scripts/pack.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+pnpm pack && npm install -g $(pwd)/$(ls ./*.tgz | head -1)