+++ templates/components/vectordbs/typescript/mongo/generate.mjs
@@ -2,12 +2,12 @@
 import * as dotenv from "dotenv";
 import {
   MongoDBAtlasVectorSearch,
-  SimpleDirectoryReader,
   VectorStoreIndex,
   storageContextFromDefaults,
 } from "llamaindex";
 import { MongoClient } from "mongodb";
-import { STORAGE_DIR, checkRequiredEnvVars } from "./shared.mjs";
+import { getDocuments } from "./loader.mjs";
+import { checkRequiredEnvVars } from "./shared.mjs";
 
 dotenv.config();
 
@@ -21,9 +21,7 @@ async function loadAndIndex() {
   const client = new MongoClient(mongoUri);
 
   // load objects from storage and convert them into LlamaIndex Document objects
-  const documents = await new SimpleDirectoryReader().loadData({
-    directoryPath: STORAGE_DIR,
-  });
+  const documents = await getDocuments();
 
   // create Atlas as a vector store
   const vectorStore = new MongoDBAtlasVectorSearch({